+++ api/CHANGELOG.md
@@ -199,6 +199,7 @@
 - Updated local downloader and runtime plugins to display "Loaded" instead of "Trusted" https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/562
 - Standardized revoke trust dialog titles across downloader and runtime plugins https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/561
 - Adjusted the supported versions dialog so it is not become too wide when opened from patch search https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/559
+- Standardized installer log filename and timestamp formatting with other exported logs https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/558
 
 
 # Bug fixes